fix: do not build debug a default seccomp policy in debug binaries
Rust 1.80.0 added a debug assertion that uses fcntl(F_GETFD) to ensure
the fd is still valid when it gets dropped, which broke debug builds of
firecracker.

This made us rethink on whether we'd want any default seccomp policy in
debug builds, and we decided that in most cases we don't need them and
in some cases they get in the way of prororyping and debugging.

This patch changes the default seccomp policy in debug builds to empty.

Signed-off-by: Riccardo Mancini <mancio@amazon.com>