test: do not run criterion benchmarks in no_block_pr
These tests have been failing consistently in our CI for as long as I
can remember. While each individual test false positive rate is not
high, the fact that we run many of them in a moltitude of combinations
means that at every CI run at least one fails.

With this change, we are disabling them in the PR - Optional step. We
will still be able to notice performance regressions from the
performance tests, so we're not losing much test coverage.
We're not completely deleting them because they can be useful for deep
dives and for verifying changes in those particular parts of the
codebase.

Signed-off-by: Riccardo Mancini <mancio@amazon.com>