feat(virtio-mem): dynamically add new kvm slots as memory is hotplugged
Whenever memory is hotplugged and `usable_size` needs to grow, plug a
new kvm memory slot covering the missing memory, in other words
`[base+usable_size, base+requested_size)`.

HACK: we need to store the last RAM address to build system tables as we
don't want to use the last virtio-mem address for that.
For now I've parked it in the Vmm object.

Signed-off-by: Riccardo Mancini <mancio@amazon.com>