hack(async): use try_access to handle guest memory properly (doesn't work!!)
THIS COMMIT DOESN'T WORK, BUT WAS MY FIRST STAB AT FIXING THE ISSUE.

When accessing guest memory, we shouldn't assume it's also contiguous on
the host and in the same GuestMemoryRegion.
However, properly handling this requires splitting the IO requests to
the io_uring, which is not trivial as we would need to add tracking of
the individual request parts.

Signed-off-by: Riccardo Mancini <mancio@amazon.com>