refactor(x86): do not rely on guest_memory.last_addr()
Replace references to guest_memory.last_addr() to use the specific
region they need.
On x86, we use it to build the memory map. We can simplify the entire
logic by just iterating over all memory regions and generating the
mappings from there.
We need to be careful to clamp the first 1MB of physical memory as it is
reserved.

This is needed for hotplugging via virtio-mem as not all of the memory
will be plugged at boot and hotpluggable memory regions should not be
exposed in the memory map.

Signed-off-by: Riccardo Mancini <mancio@amazon.com>