refactor(aarch64): do not rely on guest_memory.last_addr()
Replace references to guest_memory.last_addr() to use the specific
region they need.
On ARM, we want the last address of the guest memory to place the fdt
and initrd. Since we have a single region, just retrieve the first
region and use its address.

This is needed for hotplugging via virtio-mem as not all of the memory
will be plugged at boot and hotpluggable memory regions should not be
exposed in the memory map.

Signed-off-by: Riccardo Mancini <mancio@amazon.com>