refactor(x86): iterate over regions to populate memory map
The way we populate the PVH and E820 memmap is by redoing all the math
about the regions.
Replace that duplicated logic with an iteration over the guest memory,
filtering by Dram type (as we don't want to expose other regions we will
add in the future). Furthermore, clamp the first 1MB as it's reserved
for the system.

Signed-off-by: Riccardo Mancini <mancio@amazon.com>